Move the LoadableAvatar component + add sharp package
