+++ apps/earn-protocol-landing-page/package.json
@@ -22,12 +22,13 @@
     "dayjs": "^1.11.12",
     "include-media": "2.0.0",
     "lodash-es": "^4.17.1",
-    "mixpanel": "^0.18.0",
     "mixpanel-browser": "^2.53.0",
-    "next": "^14.2.15",
+    "mixpanel": "^0.18.0",
     "next-intl": "^3.21.1",
-    "react": "^18.3.1",
+    "next": "^14.2.15",
     "react-dom": "^18.3.1",
+    "react": "^18.3.1",
+    "sharp": "^0.33.5",
     "usehooks-ts": "^3.1.0"
   },
   "devDependencies": {
@@ -37,10 +38,10 @@
     "@types/loadable__component": "^5.13.9",
     "@types/lodash-es": "^4.17.1",
     "@types/mixpanel-browser": "^2.49.0",
-    "@types/node": "^20.14.2",
     "@types/node-fetch": "^2.6.11",
-    "@types/react": "^18.3.1",
+    "@types/node": "^20.14.2",
     "@types/react-dom": "^18.3.0",
+    "@types/react": "^18.3.1",
     "json-to-ts": "^1.7.0",
     "node-fetch": "^2.7.0"
   },
