+++ apps/earn-protocol/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "dotenv -e ../../.env -e ../../.env.local -- next dev -p 3002 --turbo",
     "build": "dotenv -e ../../.env -e ../../.env.local -- next build",
-    "start": "next start -p 3002",
+    "start": "dotenv -e ../../.env -e ../../.env.local -- next start -p 3002",
     "lint": "eslint .",
     "lint-timing": "TIMING=1 eslint ."
   },
@@ -57,6 +57,7 @@
     "react-dom": "^18.3.1",
     "react": "^18.3.1",
     "recharts": "^2.12.7",
+    "sharp": "^0.33.5",
     "usehooks-ts": "^3.1.0",
     "viem": "2.9.19",
     "wagmi": "^2.10.9"
@@ -64,18 +65,18 @@
   "devDependencies": {
     "@summerfi/eslint-config": "workspace:*",
     "@summerfi/typescript-config": "workspace:*",
+    "@types/jsonwebtoken": "^9.0.0",
     "@types/loadable__component": "^5.13.9",
     "@types/lodash-es": "^4.17.1",
     "@types/mixpanel-browser": "^2.49.0",
-    "@types/node": "^20.14.2",
     "@types/node-fetch": "^2.6.11",
-    "@types/react": "^18.3.1",
+    "@types/node": "^20.14.2",
     "@types/react-dom": "^18.3.0",
-    "@types/jsonwebtoken": "^9.0.0",
+    "@types/react": "^18.3.1",
     "json-to-ts": "^1.7.0",
     "node-fetch": "^2.7.0",
-    "tailwindcss": "^3.4.4",
-    "postcss": "^8"
+    "postcss": "^8",
+    "tailwindcss": "^3.4.4"
   },
   "packageManager": "pnpm@8.15.4"
 }
