+++ packages/app-earn-ui/src/components/atoms/GenericTokenIcon/GenericTokenIcon.tsx
@@ -1,7 +1,8 @@
 'use client'
 
 import { type FC } from 'react'
-import loadable from '@loadable/component'
+
+import { LoadableAvatar } from '@/components/atoms/LoadableAvatar/LoadableAvatar'
 
 import genericTokenIconStyles, {
   type ClassNames,
@@ -15,20 +16,6 @@ interface GenericTokenIconProps {
 }
 
 const scaleFactor = 0.8125
-
-const AvatarLoadingState = loadable(() => import('boring-avatars'), {
-  cacheKey: (props) => props.name,
-  fallback: (
-    <svg viewBox="0 0 6.35 6.35" color="inherit" display="inline-block" width={24} height={24}>
-      <circle
-        style={{ fill: '#9d9d9d', fillOpacity: 0.350168, strokeWidth: 0.340624 }}
-        cx="3.175"
-        cy="3.175"
-        r="3.175"
-      />
-    </svg>
-  ),
-})
 
 export const GenericTokenIcon: FC<GenericTokenIconProps> = ({
   variant = 'smallIcon',
@@ -47,7 +34,7 @@ export const GenericTokenIcon: FC<GenericTokenIconProps> = ({
 
   return (
     <div className={genericTokenIconStyles[variant]} style={customSize ? customSizeStyle : {}}>
-      <AvatarLoadingState
+      <LoadableAvatar
         size={innerSize}
         name={btoa(symbol)}
         variant="marble"
