+++ packages/app-earn-ui/src/index.ts
@@ -18,6 +18,7 @@ export { Expander } from './components/atoms/Expander/Expander'
 export { TableRowAccent } from './components/atoms/TableRowAccent/TableRowAccent'
 export { ToggleButton } from './components/atoms/ToggleButton/ToggleButton'
 export { Timeframes } from './components/atoms/Timeframes/Timeframes'
+export { LoadableAvatar } from './components/atoms/LoadableAvatar/LoadableAvatar'
 
 export { Footer } from './components/layout/Footer/Footer'
